restructure database integration tests
