+++ tests/Oryx.Integration.Tests/Php/PhpSqlServerIntegrationTests.cs
@@ -1,0 +1,74 @@
+﻿// --------------------------------------------------------------------------------------------
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT license.
+// --------------------------------------------------------------------------------------------
+
+using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Threading.Tasks;
+using Microsoft.Oryx.BuildScriptGenerator.Common;
+using Microsoft.Oryx.Integration.Tests.Fixtures;
+using Microsoft.Oryx.Tests.Common;
+using Xunit;
+using Xunit.Abstractions;
+
+namespace Microsoft.Oryx.Integration.Tests
+{
+    [Trait("category", "php")]
+    [Trait("db", "sqlserver")]
+    public class PhpSqlServerIntegrationTests : PlatformEndToEndTestsBase
+    {
+        private const int ContainerPort = 3000;
+        private const string DefaultStartupFilePath = "./run.sh";
+
+        public PhpSqlServerIntegrationTests(ITestOutputHelper output) : base(output, null)
+        {
+        }
+
+        [Theory]
+        [InlineData("7.3", "github-actions")]
+        [InlineData("7.3", "github-actions-buster")]
+        [InlineData("7.3", "latest")]
+        // pdo_sqlsrv only supports PHP >= 7.3
+        public async Task PhpApp_UsingPdo(string phpVersion, string imageTag)
+        {
+            // Arrange
+            var appName = "sqlsrv-example";
+            var hostDir = Path.Combine(_hostSamplesDir, "php", appName);
+            var volume = DockerVolume.CreateMirror(hostDir);
+            var appDir = volume.ContainerDir;
+            var script = new ShellScriptBuilder()
+                .AddCommand($"oryx create-script -appPath {appDir} -bindPort {ContainerPort}")
+                .AddCommand(DefaultStartupFilePath)
+                .ToString();
+
+            await EndToEndTestHelper.BuildRunAndAssertAppAsync(
+                appName,
+                _output,
+                new List<DockerVolume> { volume },
+                _imageHelper.GetBuildImage(imageTag),
+                "oryx",
+                new[] { "build", appDir, "--platform", "php", "--platform-version", phpVersion },
+                _imageHelper.GetRuntimeImage("php", phpVersion),
+                SqlServerDbTestHelper.GetEnvironmentVariables(),
+                ContainerPort,
+                "/bin/bash",
+                new[]
+                {
+                    "-c",
+                    script
+                },
+                async (hostPort) =>
+                {
+                    var data = await _httpClient.GetStringAsync($"http://localhost:{hostPort}/");
+                    Assert.Equal(
+                        DbContainerFixtureBase.GetSampleDataAsJson(),
+                        data.Trim(),
+                        ignoreLineEndingDifferences: true,
+                        ignoreWhiteSpaceDifferences: true);
+                });
+        }
+
+    }
+}
