+++ tests/Oryx.Integration.Tests/Python/PythonPostgreSqlIntegrationTests.cs
@@ -1,0 +1,36 @@
+﻿// --------------------------------------------------------------------------------------------
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT license.
+// --------------------------------------------------------------------------------------------
+
+using System.IO;
+using System.Threading.Tasks;
+using Xunit;
+using Xunit.Abstractions;
+
+namespace Microsoft.Oryx.Integration.Tests
+{
+    [Trait("category", "python")]
+    [Trait("db", "postgres")]
+    public class PythonPostgreSqlIntegrationTests : DatabaseTestsBase, IClassFixture<Fixtures.PostgreSqlDbContainerFixture>
+    {
+        public PythonPostgreSqlIntegrationTests(ITestOutputHelper output, Fixtures.PostgreSqlDbContainerFixture dbFixture)
+            : base(output, dbFixture)
+        {
+        }
+
+        [Theory]
+        [InlineData("github-actions")]
+        [InlineData("github-actions-buster")]
+        [InlineData("latest")]
+        public async Task Python37App_PostgreSqlDB(string imageTag)
+        {
+            await RunTestAsync(
+                "python",
+                "3.7",
+                Path.Combine(HostSamplesDir, "python", "postgres-sample"),
+                buildImageName: _imageHelper.GetBuildImage(imageTag));
+        }
+
+    }
+}
